Store background images on the filesystem instead of in the database
We have a new table called `images` that stores data about background
images, including its id, title, owner and whether it is active.
We use the id as its filename on disk to ensure that there are no files
with the same name. We still have to restrict access to `/static/images`
to make sure no one can see other people's background images.